requirements.txt
bert_score==0.3.13
bitsandbytes==0.45.5
huggingface_hub==0.30.2
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
Requests==2.32.3
rouge_score==0.1.2
scikit_learn==1.6.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
trl==0.15.2
unsloth_zoo==2025.4.2
vllm==0.8.5.post1
